Migrate _register_to_engine_info_bootstrap to RemoteInstanceWeightTransport
